Extract logic of how to resolve artifact urls into the dependency resolver.
